+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added emote menu.
+- Added connection status when connecting to chat.
+
 ## [0.3.12] - 2020-12-30
 
 - Added correct badges to messages.